+Added command argument parsing system. %Changed the already made commands to use this new argument parsing system.
INFO: In the constructor of each command arguments have to be defined.
I've added quite a lot already but more will be added over time like entity types, enchants etc.
Arguments can be OPTIONAL, REQUIRED or REQUIRED_CONSOLE.
For the last one a player doesn't need to pass in the argument but the console does.
If something is required and invalid it will send the CommandSender an error message.
Many arguments have multiple constructors for validation like StringArgument can check for a matching string, the amount of characters or even if a string starts and/or ends with a certain string.
See the already made commands for examples.