+Command options +Command modifiers +Optional command arguments %Modified HealCmd to work with the new stuff.
INFO: Alright so how this works is that each command can add modifiers, command options and optional arguments.
All command options will be added as optional argument so it's not needed to add those.
Modifiers are things that modify the command behavior but don't have any user input. For example -s doesn't send messages to the user and so on.
Command options are defined in the CommandOptions config which are mostly things like booleans, integers and strings.
For example if the heal command should remove potion effects or not is a perfect command option.
But something like the name for an item in /give would be terrible since it would put that name on each item you give.
So command options are like optional arguments but have configurable defaults.
Then last but not like the optional arguments are arguments that the player can add to the command for example name:Some_name for give an item a name.
Optional arguments should be always based on the user input and have no defaults.

How it works:
Inside ArgumentParseResults all the modifiers and optional args are stored that the user passed in the command.
So, inside a command you can do result.hasModifier("-s") and this will return true if one of the arguments was "-s"
You can also do result.hasOptionalArg("name") and result.getOptionalArg("name")
The getOptionalArg needs to be casted tot he OptionType that was given so if it's a BoolOption you cast it to a Boolean
See HealCmd for an example how it all comes together.
Make sure that after you parse the arguments and get the result back that you set the command arguments to the parsed arguments.
This will remove all modifiers and optional args so args = result.getArgs();