+Added modules config to enable/disable modules. -Removed enabled and usage lines from Commands config. %Cleanup command registration (failed conflict merge messed it up)
INFO: The reason for modules is so commands can be grouped together.
For example warps it would be weird if you could enable /delwarp but have /warp disabled for example.
So now commands can be bundled together in modules and enabled/disabled in groups.
Also means commands can be forced enabled by not specifying a module for example for the /essence command.
Modules will also be uses for other things that aren't commands.