+Item aliases stuff +Iteminfo command.
Note: Uncompleted because I'm moving today I won't have my PC for a bit and I will continue on it on my laptop.
Still need to fill in default aliases, dot eh iteminfo command properly and so on.
Also need to do a smilar system (less advanced) for other enums.