%Massive redesign to the ItemParser.
Info: You need to create an instance of the ItemParser now and pass in a string or item.
You can then check if the parse result was valid, if not you can get the error message if successful you can get the item or string.
See TestCmd for some examples how the ItemParser works.

Note: It's still not completed like I still need to do parsing items to strings.
Also need to do enchants etc but the alias config needs to be done for those things first.
And the message system needs to be redesigned to complete this.