%Complete redesign to message system.
INFO: The Message enum now creates EMessage instances.
This means that other plugins will be able to create EMessages too.
To send a message you would do Message.KEY.msg() to get the EMessage.
Then you can do Message.KEY.msg().getMsg(true) to get the config value and format it so it's also a bit easier to get messages.
The only downside to this system is that there is no way to clean old messages from the config but it's not a big deal.

NOTE: Still need to convert everything inside Essence to EMessage instead of Message.
For example the ItemParser, constructors of modifiers, options, arguments etc.