+Database class to establish connections and such which can  be overwritten by driver specific classes. +Database classes for MySql and SqlLite drivers. +Establish database connection on enable based on the settings in DataStorage.yml +Added a config option if the plugin should be disabled when there is no database connection on startup. -Removed debug messages for DataStorage
Info: The database classes are very basic for now and only support update queries and regular queries.
Everything else can be done by just getting the connection the two methods in Database are just wrappers to make it a little easier.
To use prepared statements etc you just have to do it directly with the connection.
Might add more support for things in the Database later but I rather don't limit the usage too much.
No tables or anything are created yet in this commit it's just connections to the database.

Note: It will be important that all things related to data storage will check database connections first.
So for example /ban will have to be disabled when there is no database connection and so on.