%Refactor player data storage for modules completely. +Added getModule() method in Modules.java to get a module class instance. %Renamed DataType to DataModules
Info: Data is no longer stored per player.
Instead it's all stored in the module class inside a map or it can be anything. (See BanModule)
This makes it a lot easier to manage data for modules.
The only downside is that it takes a couple extra lines to get/set data from a player but we'll probably add some wrapper methods to make this easier.