+Added the EntityParser to parse strings to EEntity's (see info/note below) +Added a enum with EntityTag which contains all the tags that can be used in entity parsing. +Added entity argument to test command for testing entity spawning.
Note: The parser is not 100% completed.
Still need to do the parsing from EEntity to string. The current version can only parse strings to EEntity.
Also need to do something with locations so that the location can be stored in a entity string and spawn it at that location.
And potion effects need to be manually parsed using aliases.
Lastly it needs to fetch the entity type from aliases but the aliases for that haven't been implemented.

Info: The parser is completely based on the EntityTag enum.
So it's really easy to add/remove/modify tags to the parser without actually touching the parser.
Most of the code in the parser is for the syntax since it's really advanced how the syntax works. (See example below)
There are a few tag specific things in the parser.
For example ListArg and MappedListArg return a string so those have to be convert before calling the method (which uses reflection btw)
And some methods need custom calls like if you set health it needs to set the max health if the health is more than the max health etc.

This is the basic syntax:
blaze(health:1)>spider(health:200)>armorstand(marker:true,name:"&a&lExample > test",namevisible:true,helmet:"humanhead 1 player:Worstboy",baseplate:false)
So each entity is seperated with a > and they get stacked ontop of eachother.
It can also be a - and then they can leashed together.
Then all entity data needs to be between ().
Each data kv pair is seperated by a comma.
When the data is a string with spaces it can be put in quotes.
It also supports escaping characters with a backslash \>\-\(\) etc