+Added LocationArg.
(Sorry war but made it from scratch as your method was still missing some stuff and it was quite messy)
Will have to redo LocationArgument to do it similar to this one.
However, I'll probably make something so that args can share the same parsing implementions for Arg and Argument.