+Added advanced command linkage system.
Note: Still need to implement ConflictLink and LinkLink but I have to refactor arguments before I can continue so I'll commit this first.
Don't revert to this commit as commands are completely broken in this version.

Info: Links can be set between optional args, modifiers and command arguments.
For example the heal command now has a link between the -a modifier and the player argument.
When -a is specified it will remove the player argument making the syntax /heal -a [max]
So the first value for the link is the value that needs to be specified for the link to work and the second value is the value that gets modified.
This value is a string containing the name of a modifier, optional arg or argument. in this case ("-a", "player")
There are several link options and I'll explain them in depth on the wiki but it should be pretty straight forward.